Added textures and font rendering, improved split and cull hueristic
By adding textures I was able to create a scene that shows the moon (1
unit = 1km).
I added font rendering and fixed a bunch of geometry shader loading
problems in the process.
The culling works a bit better now, but still needs one more fix. The
splitting is now based on the apparent screen size of the triangles so
they always split if they reach a set pixel size on the screen.